TO DO: Server's information, Check telegram-channel, change password

+++ app/Http/Controllers/ServersController.php
@@ -13,9 +13,13 @@ use Illuminate\Support\Facades\Validator;
 class ServersController extends Controller
 {
     //
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index(Request $request)
     {
-        // print_r($request->session()->all());
         return view('servers.servers');
     }
 
@@ -48,7 +52,13 @@ class ServersController extends Controller
 
         $data = [];
         $favorite = $request->input('favorite');
+        $status = $request->input('status');
+        $users = $request->input('users');
+
+
         if(!empty($favorite)) $data['favorite'] = '1';
+        if($status != "") $data['status'] = $status;
+        if(!empty($users) && $user->isAdmin()) $data['users'] = $users;
         
         if(!$user->isAdmin()) {
            $data['userId'] = Auth::id();
@@ -60,12 +70,26 @@ class ServersController extends Controller
 
     public function edit($id)
     {
+        $user = new User();
+        $server = new Server();
+
+        if(!$user->isAdmin()) {
+            $isLink = $server->isLink($id);
+            if(!$isLink) return view('404');
+        }
         $serverData = Server::findOrFail($id);
         return view('servers.manage',$serverData);
     }
 
     public function update($id, User $user)
     {
+        $user = new User();
+        $server = new Server();
+
+        if(!$user->isAdmin()) {
+            $isLink = $server->isLink($id);
+            if(!$isLink) return view('404');
+        }
         $data = request()->validate([
             'name' => ['required', 'string', 'max:255'],
             'host' => ['required', 'string', 'max:255'],
@@ -98,6 +122,13 @@ class ServersController extends Controller
 
     public function destroy($id, Server $server)
     {
+        $user = new User();
+        $server = new Server();
+
+        if(!$user->isAdmin()) {
+            $isLink = $server->isLink($id);
+            if(!$isLink) return view('404');
+        }
         $userType = Auth()->user()->type;
         
         $result = $server->deleteServer($id,$userType);
